Conflicts:
	DongnaeHanbaki/src/main/java/com/spring/dongnae/home/controller/HomeController.java
	DongnaeHanbaki/src/main/java/com/spring/dongnae/user/controller/UserController.java
	DongnaeHanbaki/src/main/java/com/spring/dongnae/utils/auth/GetAuthenticInfo.java
	DongnaeHanbaki/src/main/webapp/WEB-INF/patials/commonBody.jsp
	DongnaeHanbaki/src/main/webapp/WEB-INF/views/home/home.jsp
	DongnaeHanbaki/src/main/webapp/resources/js/sidebar/moim.js

+++ DongnaeHanbaki/src/main/java/com/spring/dongnae/socket/service/MoimService.java
@@ -57,14 +57,30 @@ public class MoimService {
         return moim;
     }
     
-    public Moim addParticipantToMoim(String moimId, String token) {
-    	Optional<Moim> moimOptional = moimRepository.findById(moimId);
+    public Moim addParticipantToMoim(String moimName, String token) {
+    	Optional<Moim> moimOptional = moimRepository.findByName(moimName);
     	Optional<UserRooms> userRoomsOptional = userRoomsRepository.findById(token);
     	if (moimOptional.isPresent() && userRoomsOptional.isPresent()) {
     		Moim moim = moimOptional.get();
     		UserRooms userRooms = userRoomsOptional.get();
     		moim.addParticipant(userRooms);
+    		chatRoomService.addUserToChatRoom(moim.getChatRoomId(), userRooms);
     		userRooms.addMoim(moim);
+    		userRoomsRepository.save(userRooms);
+    		return moimRepository.save(moim);
+    	}
+    	return null;
+    }
+    
+    public Moim banUserAtMoim(String moimName, String token) {
+    	Optional<Moim> moimOptional = moimRepository.findByName(moimName);
+    	Optional<UserRooms> userRoomsOptional = userRoomsRepository.findById(token);
+    	if (moimOptional.isPresent() && userRoomsOptional.isPresent()) {
+    		Moim moim = moimOptional.get();
+    		UserRooms userRooms = userRoomsOptional.get();
+    		moim.banUser(userRooms);
+    		userRooms.removeMoim(moim);
+    		chatRoomService.banUserToChatRoom(moim.getChatRoomId(), userRooms);
     		userRoomsRepository.save(userRooms);
     		return moimRepository.save(moim);
     	}
@@ -86,8 +102,12 @@ public class MoimService {
     	Optional<Board> boardOptional = boardRepository.findById(boardId);
     	if (boardOptional.isPresent()) {
     		Board board = boardOptional.get();
-    		boardRepository.delete(board);
-    		return true;
+    		if (board.getAuthor().equals(getAuthenticInfo.GetToken()) || board.getMoim().getLeader().equals(getAuthenticInfo.GetToken()) || board.getMoim().getSubLeader().contains(getAuthenticInfo.GetToken())) {    			
+    			boardRepository.delete(board);
+    			return true;
+    		} else {
+    			return false;
+    		}
     	}
     	return false;
     }
@@ -150,14 +170,10 @@ public class MoimService {
     public Optional<Board> getBoardById(String boardId) {
         return boardRepository.findById(boardId);
     }
-//    
-//    public List<Moim> searchMoimByTitle(String query) {
-//        Pageable pageable = PageRequest.(0, 5); // 첫 페이지에서 5개의 결과만 가져옴
-//        return moimRepository.findByNameContainingIgnoreCase(query, pageable).getContent();
-//    }
     
-//    public List<Moim> getMoimByToken(String token) {
-//    	Optional<UserRooms> userRoomsOptional = userRoomsRepository.findById(token);
-//    	return userRoomsOptional.map(UserRooms::getMoims).orElse(null);
-//    }
+    public Page<Moim> searchMoims(String name, List<String> excludeIds) {
+    	Pageable pageable = new PageRequest(0, 5);  // 첫 페이지에서 최대 5개의 결과
+        return moimRepository.findByNameContainingIgnoreCaseAndIdNotIn(name, excludeIds, pageable);
+    }
+
 }
